requirements.txt
# Core ML and Computer Vision
ultralytics>=8.0.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.5.0
numpy>=1.24.0

# Web Framework and API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
jinja2>=3.1.0

# Audio and Alarm System
pygame>=2.5.0
playsound>=1.3.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
loguru>=0.7.0
psutil>=5.9.0

# Optional: For advanced features
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
